Improved transaction code and started writing blockchain.
I simplified the output of the transaction code, using hashes of previous
transactions instead of the full text of all previous transactions.

I also started implementing the blockchain. I just Clojure's (Java's?)
hash function instead of SHA-256 but the principles remain. I still have to
implement the block verification that is so important to the blockchain's
function, and I am not sure how I will implement and test the longest-chain
behaviors of the paper.